@@ -5,7 +5,7 @@ import {AppComponent} from './components/app.component';
 import {NavbarComponent} from './components/include/navbar/navbar.component';
 import {AppRoutingModule, routingComponents} from "./app-routing.module";
 import {FormsModule} from "@angular/forms";
-import {HttpClientModule} from "@angular/common/http";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 import {FooterComponent} from './components/include/footer/footer.component';
 import {CVComponent} from './components/include/cv/cv.component';
 import {CertifComponent} from './components/include/cv/certif/certif.component';
@@ -25,36 +25,30 @@ import {
 } from "./components/include/skeletons/remove-ip-skeletons/remove-ip-skeletons.component";
 import {AddIpSkeletonsComponent} from "./components/include/skeletons/add-ip-skeletons/add-ip-skeletons.component";
 
-@NgModule({
-  declarations: [
-    routingComponents,
-    AppComponent,
-    NavbarComponent,
-    FooterComponent,
-    CertifComponent,
-    ExperienceComponent,
-    LanguagesComponent,
-    CompetencesComponent,
-    CVComponent,
-    ServicesComponent,
-    TempladminComponent,
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    HttpClientModule,
-    FormsModule,
-    NgxSonnerToaster,
-    NgxTurnstileModule,
-    WhitelistedIpsSkeletonsComponent,
-    RemoveIpSkeletonsComponent,
-    AddIpSkeletonsComponent,
-    ContactSkeletonsComponent
-  ],
-  providers: [
-    Title
-  ],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        routingComponents,
+        AppComponent,
+        NavbarComponent,
+        FooterComponent,
+        CertifComponent,
+        ExperienceComponent,
+        LanguagesComponent,
+        CompetencesComponent,
+        CVComponent,
+        ServicesComponent,
+        TempladminComponent,
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        AppRoutingModule,
+        FormsModule,
+        NgxSonnerToaster,
+        NgxTurnstileModule,
+        WhitelistedIpsSkeletonsComponent,
+        RemoveIpSkeletonsComponent,
+        AddIpSkeletonsComponent,
+        ContactSkeletonsComponent], providers: [
+        Title,
+        provideHttpClient(withInterceptorsFromDi())
+    ] })
 export class AppModule {
 }
